server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.196
transformers==4.35.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
motor==3.3.1 